Added a local best scores tracking
There is now an option to keep your score, along with your name, and get it onto the best score list. If you have surpassed one of the top 3 scores, you will get on the table.
In addition, I have also fixed some bugs and issues regarding the responsiveness of the page.